usercircle 업데이트 권한 설정

+++ sharkle/circle/views.py
@@ -211,6 +211,12 @@ class UserCircleViewSet(viewsets.GenericViewSet):
 class UserCircleUpdateSet(viewsets.GenericViewSet):
     serializer_class = CircleSerializer
     permission_classes = (permissions.AllowAny,)  # 테스트용 임시
+    no_authority_error = ExceptionResponse(
+        status=status.HTTP_400_BAD_REQUEST,
+        detail="권한이 부족합니다",
+        code=ErrorCode.NOT_ALLOWED,
+    ).to_response()
+
     # GET /circle/{id}/account/{id}/
     def list(self, request, circle_id, user_id):
 
@@ -224,42 +230,6 @@ class UserCircleUpdateSet(viewsets.GenericViewSet):
 
         return Response(status=status.HTTP_200_OK, data=user_status(circle, user))
 
-    # GET /circle/{id}/account/{id}/{name}/
-    def retrieve(self, request, circle_id, user_id, pk):
-
-        error, circle = find_circle(circle_id)
-        if error:
-            return error
-
-        error, user = find_user(user_id, request.user)
-        if error:
-            return error
-
-        if pk == "alarm":
-            return Response(
-                status=status.HTTP_200_OK,
-                data=bool(
-                    UserCircle_Alarm.objects.get_or_none(user=user, circle=circle)
-                ),
-            )
-        if pk == "member":
-            return Response(
-                status=status.HTTP_200_OK,
-                data=bool(
-                    UserCircle_Member.objects.get_or_none(user=user, circle=circle)
-                ),
-            )
-        if pk == "manager":
-            return Response(
-                status=status.HTTP_200_OK,
-                data=bool(
-                    UserCircle_Member.objects.get_or_none(
-                        user=user, circle=circle, is_manager=True
-                    )
-                ),
-            )
-        return Response(status=status.HTTP_200_OK)
-
     # PUT /circle/{id}/account/{id}/{name}/
     def update(self, request, circle_id, user_id, pk):
 
@@ -271,18 +241,33 @@ class UserCircleUpdateSet(viewsets.GenericViewSet):
         if error:
             return error
 
+        my = request.user
+        my_membership = user_membership(circle, request.user)
+
         if pk == "alarm":
-            UserCircle_Alarm.objects.get_or_create(user=user, circle=circle)
+            # 내 계정의 알림만 변경 가능
+            if user == my:
+                UserCircle_Alarm.objects.get_or_create(user=user, circle=circle)
+            else:
+                return self.no_authority_error
 
         if pk == "member":
-            UserCircle_Member.objects.get_or_create(user=user, circle=circle)
+            # 'make_new_member' 권한 이상의 유저는 member 권한 부여 가능
+            if my_membership[1] >= circle.make_new_member:
+                UserCircle_Member.objects.get_or_create(user=user, circle=circle)
+            else:
+                return self.no_authority_error
 
         if pk == "manager":
-            user_circle_member = UserCircle_Member.objects.get_or_create(
-                user=user, circle=circle
-            )[0]
-            user_circle_member.is_manager = True
-            user_circle_member.save()
+            # 관리자는 manager 권한 부여 가능
+            if my_membership[0] == "관리자":
+                user_circle_member = UserCircle_Member.objects.get_or_create(
+                    user=user, circle=circle
+                )[0]
+                user_circle_member.is_manager = True
+                user_circle_member.save()
+            else:
+                return self.no_authority_error
 
         return Response(
             status=status.HTTP_200_OK,
@@ -300,24 +285,34 @@ class UserCircleUpdateSet(viewsets.GenericViewSet):
         if error:
             return error
 
+        my = request.user
+        my_membership = user_membership(circle, request.user)
+
         if pk == "alarm":
-            if user_circle_alarm := UserCircle_Alarm.objects.get_or_none(
-                user=user, circle=circle
-            ):
-                user_circle_alarm.delete()
+            # 내 계정의 알람만 변경 가능
+            if user == my:
+                if user_circle_alarm := UserCircle_Alarm.objects.get_or_none(user=user, circle=circle):
+                    user_circle_alarm.delete()
+            else:
+                return self.no_authority_error
 
         if pk == "member":
-            if user_circle_member := UserCircle_Member.objects.get_or_none(
-                user=user, circle=circle
-            ):
-                user_circle_member.delete()
+            # 내 계정 혹은 관리자 계정이 member 권한 삭제 가능
+            if my == user or my_membership[0] == "관리자":
+
+                if user_circle_member := UserCircle_Member.objects.get_or_none(user=user, circle=circle):
+                    user_circle_member.delete()
+            else:
+                return self.no_authority_error
 
         if pk == "manager":
-            if user_circle_member := UserCircle_Member.objects.get_or_none(
-                user=user, circle=circle
-            ):
-                user_circle_member.is_manager = False
-                user_circle_member.save()
+            # 내 계정 혹은 관리자 계정이 manager 권한 삭제 가능
+            if my == user or my_membership[0] == "관리자":
+                if user_circle_member := UserCircle_Member.objects.get_or_none(user=user, circle=circle):
+                    user_circle_member.is_manager = False
+                    user_circle_member.save()
+            else:
+                return self.no_authority_error
 
         return Response(
             status=status.HTTP_200_OK,
